get user authenticated

+++ src/main/java/com/uade/grupo4/backend_ecommerce/controller/dto/UserDto.java
@@ -4,14 +4,14 @@ import lombok.Data;
 
 @Data
 public class UserDto {
-    private int id;
+    private Long id;
     private String username;
     private String email;
     private String password;
     private String firstName;
     private String lastName;
 
-    public UserDto(int id, String username, String email, String firstName, String lastName) {
+    public UserDto(Long id, String username, String email, String firstName, String lastName) {
         this.id = id;
         this.username = username;
         this.email = email;
